fix: migrate training stack to PyTorch Lightning 2.x on GPU
- timm.layers imports; torch.load with weights_only=False for ckpt
- normalize_precision, resolve_accelerator, PL2 Trainer strategy helpers
- Default precision 16-mixed and accelerator gpu for CUDA hosts
- Remove PL 1.x code paths in run, predict, test, validation

+++ moftransformer/config_ex.py
@@ -97,7 +97,7 @@ def config():
         load_path = ""
 
     num_workers = 16  # the number of cpu's core
-    precision = 16
+    precision = "16-mixed"
 
     # normalization target
     mean = None
